+++ lint-staged.config.js
@@ -1,0 +1,4 @@
+export default {
+  '{src,test}/**/*.js': 'eslint --fix',
+  '{src,docs}/**/*.css': 'stylelint --fix',
+};