+++ test/helpers.js
@@ -1,13 +1,2 @@
-window.loadImageAsBlob = (url, done) => {
-  const xhr = new XMLHttpRequest();
-
-  xhr.onload = () => {
-    const blob = xhr.response;
-
-    blob.name = url.replace(/^.*?(\w+\.\w+)$/, '$1');
-    done(blob);
-  };
-  xhr.open('GET', url);
-  xhr.responseType = 'blob';
-  xhr.send();
-};
+// Re-export from setup for backwards compatibility
+export { loadImageAsBlob, Compressor, utilities } from './setup.js';